+++ ipc_sun_sync/parser.py
@@ -1,4 +1,5 @@
 import argparse
+from datetime import timedelta
 import pathlib
 import sys
 import logging
@@ -7,6 +8,7 @@ from typing import Dict
 import astral
 import pytz
 import yaml
+import pytimeparse
 
 from .constants import Config, ConfigMethod, ConfigIPC
 from . import __description__
@@ -14,6 +16,13 @@ from . import __description__
 
 def parse_args():
     parser = argparse.ArgumentParser(description=__description__)
+    parser.add_argument(
+        "-v",
+        "--verbose",
+        dest="verbose",
+        action="store_true",
+        help="enable verbose logging",
+    )
     parser.add_argument(
         "--check", action="store_true", dest="check", help="check all ipc"
     )
@@ -28,13 +37,6 @@ def parse_args():
         and "-T" not in sys.argv
         and "--timezones" not in sys.argv,
         help="configuration file path",
-    )
-    parser.add_argument(
-        "-v",
-        "--verbose",
-        dest="verbose",
-        action="store_true",
-        help="enable verbose logging",
     )
     parser.add_argument(
         "-V",
@@ -67,15 +69,16 @@ def parse_yml_or_exit(path: pathlib.Path):
 
 
 def parse_method_or_exit(yml: Dict) -> ConfigMethod:
-    if "method" not in yml:
-        return ConfigMethod.CGI
-    if yml["method"] == "cgi":
+    if "method" not in yml or yml["method"] == "cgi":
         return ConfigMethod.CGI
     if yml["method"] == "rpc":
         return ConfigMethod.RPC
-
     logging.error("method invalid")
     exit(1)
+
+
+def parse_offset(offset: str) -> timedelta:
+    return timedelta(seconds=pytimeparse.parse(offset))
 
 
 def parse_config_or_exit(yml: Dict) -> Config:
@@ -84,9 +87,17 @@ def parse_config_or_exit(yml: Dict) -> Config:
         logging.error("Timezone '%s' is invalid", yml["timezone"])
         exit(1)
 
+    username = str(yml["username"]) if "username" in yml else "admin"
+    password = str(yml["password"])
+    method = parse_method_or_exit(yml)
+    sunrise_offset = (
+        parse_offset(yml["sunrise_offset"]) if "sunrise_offset" in yml else timedelta()
+    )
+    sunset_offset = (
+        parse_offset(yml["sunset_offset"]) if "sunset_offset" in yml else timedelta()
+    )
+
     config: Config = {
-        "username": str(yml["username"]) if "username" in yml else "admin",
-        "password": str(yml["password"]),
         "location": astral.LocationInfo(
             name="custom",
             region="custom",
@@ -94,7 +105,6 @@ def parse_config_or_exit(yml: Dict) -> Config:
             latitude=float(yml["latitude"]),
             longitude=float(yml["longitude"]),
         ),
-        "method": parse_method_or_exit(yml),
         "ipc": [],
     }
 
@@ -102,12 +112,16 @@ def parse_config_or_exit(yml: Dict) -> Config:
         ipc_config: ConfigIPC = {
             "ip": str(c["ip"]),
             "name": str(c["name"]) if "name" in c else str(c["ip"]),
-            "username": str(c["username"]) if "username" in c else config["username"],
-            "password": str(c["password"]) if "password" in c else config["password"],
+            "username": str(c["username"]) if "username" in c else username,
+            "password": str(c["password"]) if "password" in c else password,
             "channel": int(c["channel"]) if "channel" in c else 0,
-            "method": parse_method_or_exit(c["method"])
-            if "method" in c
-            else config["method"],
+            "method": parse_method_or_exit(c["method"]) if "method" in c else method,
+            "sunrise_offset": parse_offset(c["sunrise_offset"])
+            if "sunrise_offset" in c
+            else sunrise_offset,
+            "sunset_offset": parse_offset(c["sunset_offset"])
+            if "sunset_offset" in c
+            else sunset_offset,
         }
 
         config["ipc"].append(ipc_config)